orbital-governance-mvp/requirements.txt
flask
flask-cors
psycopg2-binary
skyfield
sqlalchemy
geoalchemy2
shapely
numpy
